Move trait test to separate directory
