+++ e2e/common/traits/affinity_test.go
@@ -19,7 +19,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package common
+package traits
 
 import (
 	"fmt"
@@ -47,7 +47,7 @@ func TestAffinityTrait(t *testing.T) {
 		if hostname != "" {
 			t.Run("Run Java with node affinity", func(t *testing.T) {
 				RegisterTestingT(t)
-				Expect(Kamel("run", "-n", ns, "files/Java.java",
+				Expect(Kamel("run", "-n", ns, "../files/Java.java",
 					"--name", "java1",
 					"-t", "affinity.enabled=true",
 					"-t", fmt.Sprintf("affinity.node-affinity-labels=kubernetes.io/hostname in(%s)", hostname)).Execute()).Should(BeNil())
@@ -68,7 +68,7 @@ func TestAffinityTrait(t *testing.T) {
 
 		t.Run("Run Java with pod affinity", func(t *testing.T) {
 			RegisterTestingT(t)
-			Expect(Kamel("run", "-n", ns, "files/Java.java",
+			Expect(Kamel("run", "-n", ns, "../files/Java.java",
 				"--name", "java2",
 				"-t", "affinity.enabled=true",
 				"-t", "affinity.pod-affinity-labels=camel.apache.org/integration").Execute()).Should(BeNil())
@@ -90,7 +90,7 @@ func TestAffinityTrait(t *testing.T) {
 		t.Run("Run Java with pod anti affinity", func(t *testing.T) {
 			RegisterTestingT(t)
 
-			Expect(Kamel("run", "-n", ns, "files/Java.java",
+			Expect(Kamel("run", "-n", ns, "../files/Java.java",
 				"--name", "java3",
 				"-t", "affinity.enabled=true",
 				"-t", "affinity.pod-anti-affinity-labels=camel.apache.org/integration").Execute()).Should(BeNil())
