+++ e2e/common/traits/prometheus_test.go
@@ -19,7 +19,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package common
+package traits
 
 import (
 	"fmt"
@@ -48,7 +48,7 @@ func TestPrometheusTrait(t *testing.T) {
 
 		Expect(Kamel("install", "-n", ns).Execute()).Should(BeNil())
 
-		Expect(Kamel("run", "-n", ns, "files/Java.java",
+		Expect(Kamel("run", "-n", ns, "../files/Java.java",
 			"-t", "prometheus.enabled=true",
 			"-t", fmt.Sprintf("prometheus.service-monitor=%v", createServiceMonitor)).Execute()).Should(BeNil())
 		Eventually(IntegrationPodPhase(ns, "java"), TestTimeoutLong).Should(Equal(v1.PodRunning))
